+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Added:
 - Added spawner mob xp config option
 ### Fixed:
-- 
+- Fixed armor switching
 ### Changed:
 -